Prepare pal_demo for compiling in 64-bit

+++ hypapps/trustvisor/pal_demo/pal.c
@@ -2,7 +2,7 @@
 
 void begin_pal_c() {}
 
-int my_pal(uint32_t arg1, uint32_t *arg2) {
+unsigned long my_pal(unsigned long arg1, unsigned long *arg2) {
 	return arg1 + ((*arg2)++);
 }
