+++ hypapps/trustvisor/pal_demo/pal.h
@@ -1,5 +1,5 @@
 #include <stdint.h>
 
 extern void begin_pal_c();
-extern int my_pal(uint32_t arg1, uint32_t *arg2);
+extern unsigned long my_pal(unsigned long arg1, unsigned long *arg2);
 extern void end_pal_c();
