+++ hypapps/trustvisor/pal_demo/test.c
@@ -4,11 +4,11 @@
 #include "trustvisor.h"
 
 int main(int argc, char *argv[]) {
-	unsigned int eax, result;
+	unsigned long eax, result;
 	assert(argc > 1);
-	assert(sscanf(argv[1], "%u", &eax) == 1);
+	assert(sscanf(argv[1], "%lu", &eax) == 1);
 	result = vmcall(TV_HC_TEST, eax, 0, 0, 0);
-	printf("%u = vmcall(TV_HC_TEST, %u)\n", result, eax);
+	printf("%lu = vmcall(TV_HC_TEST, %lu)\n", result, eax);
 	return (int)result;
 }
