+++ hypapps/trustvisor/pal_demo/vmcall.h
@@ -1,7 +1,7 @@
 #include <stdint.h>
 
-static inline uint32_t vmcall(uint32_t eax, uint32_t ecx, uint32_t edx,
-								uint32_t esi, uint32_t edi) {
+static inline uintptr_t vmcall(uintptr_t eax, uintptr_t ecx, uintptr_t edx,
+								uintptr_t esi, uintptr_t edi) {
 	asm volatile ("vmcall\n\t" : "=a"(eax) : "a"(eax), "c"(ecx), "d"(edx),
 					"S"(esi), "D"(edi));
 	return eax;
